Fix for '--ns-cert-type server is deprecated' message.
- Added extended key usage based on RFC3280 TLS rules for OpenVPNs OpenSSL configuration,
so '--remote-cert-tls' can be used instead of the old and deprecated '--ns-cert-type'
if the host certificate are newely generated with this options.
Nevertheless both directives (old and new) will work also with old CAs.

- Automatic detection in ovpnmain.cgi if the host certificate uses the new options.
If it does, '--remote-cert-tls server' will be automatically set into the client
configurations in N2N and RW client configuration files.

If it does NOT, the old '--ns-cert-type server' directive will be set in the client
configuration files for N2N and RW.